*Added an NVDAThreads module which keeps track of generator objects and has a pump function which can call next() on all of them. *Core.py now calls NVDAThreads.pump() on each iteration of its main loop *Audio module now has a synthLoop function which is executed with NVDAThreads, and this function reads commands from a queue and sends them to the current synthesizer. All normal audio functions now simply just add commands to the queue. *Added a getLastIndex function to all the synthDrivers and also one to the audio module. All the speak functions also take an index keyword argument now. *Added a sayAll script and sayAllGenerator function to NVDAObject_edit.
In summary: Audio functions can now be called from other threads such as key down/up events in keyboardHandler etc, also synths can speakText but also communicate when that certain peace of text is actually being spoken with an index.
Say all uses the current position as the index that it passes to audio.speakText.
It has a loop which does speakText, and then each time checks the index on the synth, and updates the caret position.

*Updated documentation