There are not too many changes here that will be noticeable to the end user, however they are all very important and in the end will allow NVDA to work with other technologies than just MSAA, and possibly there could be a chance that NVDA could now at some point be tinkered with to work on windows 98.
This revision Completely makes the core workings of NVDA independent of any accessibility or operating system interaction APIs.
NVDA still uses MSAA by default, but it is now possible to write NVDA objects and virtual buffers (and of course event handlers) for other APIs (such as the Java Access Bridge) with out having to tamper with MSAA support.

Focus and forground objects no longer have to be set with an MSAA event locator (window,objectID,childID), they can just directly be set with an NVDAObject. This is made possible because NVDAObject_MSAA objects now are initialised with an optional origEventLocator parameter which is the (window,objectID,childID) that the MSAA event was for. So any compairisons that need to be made against locators on MSAA objects can still be done.

the MSAA module in NVDAObjects now manages its own class map (like manager used to). Because each api (MSAA etc) has different attributes to check when choosing an NVDAObject class, class mapping has to be handled by each API with in NVDAObjects.

Virtual Buffers are pretty much the same, there is now an MSAA sub-module which manages all the MSAA virtual buffers, there is still  one function in virtualBuffers/__init__.py though, which is getVirtualBuffer, however this just checks the type of the NVDAObject being passed to it and then calls the getVirtualBuffer for the specific API (in this case, its only MSAA at the moment).

api.executeEvent has now been moved to MSAAHandler.py since it was only MSAA specific anyway.

setForegroundObjectByLocator and setFocusObjectByLocator have been removed, setFocusObject and setForegroundObject have been added.

properties specific to NVDAObject_windows now have 'window' prepended to the name (example: windowclassName, windowControlID) and NVDAObject_MSAA (MSAAChildID).

More NVDAObject docstrings have been updated.

navigator objects are now set in event_gainFocus and event_foreground in the base NVDAObject, rather than in event handlers such as MSAA. This fixes the problem with lists where you would end up on an unknown object sometimes.