Revision 230 fixes a bug where if you are running NVDA and you start internet explorer, the page will never load. If you had previously started internet explorer before starting NVDA this wouldn't have been a problem.
NVDA needs to get the events for the main document at least once, even if internet explorer says its not complete, or else NVDA will never be notified that the first page is complete.

Also virtualBuffer pass-through mode is now turned on and off at appropriate times. It is automatically turned on if you give focus to an internet explorer pane that has an editable  document. It is turned off if you give focus to anything inside an internet explorer virtual buffer (as long as the content is not editable).