reated queueHandler.py, which now handles the function queues / generator objects. core.py uses queueHandler rather than managing its own queues etc, plus this means that other modules such as keyUtils can just import queueHandler to find out about waiting keys, not core. Also this means that GUI events now do queueHandler.queueFunction rather than core.executeFunction. *Moved isKeyWating in to keyUtils.py. Changed textBuffer.py to use isKeyWaiting from keyUtils. *Created languageHandler.py, which now contains setLanguage and getAvailableLanguages (which used to be in core).
There seems to be a bit of a bug where if you hold down an arrow key with a textBuffer script, most of the time it will eventually speak when you let go, but occasionally it doesn't. Either I have stuffed up the queue logic with order of speech and scripts, or the actual queue for scripts, that isKeyWaiting checks, may not become empty for a tiny bit longer after actually getting the script out of the queue.