Add a selectable state and map it for IAccessible. Move the determination of which positive and negative states to speak from simple dicts into functions: processPositiveStates() and processNegativeStates(). The dicts only allowed decisions to be made based on the role, whereas the functions can check both role and individual states. This was done so that list items without the selectable state no longer speak "not selected". If the selectable state isn't set, the selected state isn't valid. Fixes #23.
Used this new functionality to silence the removal of the selection state on items which do not have the focused state. This stops "not selected" from being spoken in broken lists such as the page selection list in the options dialog of Mozilla Firefox/Thunderbird 3 where the state change occurs before the focus event is fired.