ntroducing NVDAControler
NVDAControler is a way that remote processes can control NVDA in various ways.
It is an RPC server/client implementation, though clients only need to call basic dll functions with out having to know about RPC etc.

NVDAControler so far supports retreaval of of NVDA's current version, and the execution of special "external" appModule events on the appModule that represents the calling process.

So for instance, if an application wanted to control NVDA to make the application more accessible in some way, it would firstly need to provide an appModule in NVDA's appModule directory, which contained a few event_external_* methods. Then, from with in the application, the executeAppModuleEvent  function could be called to execute that event.
There is no way for an appModule event to return information back to the client, but if there are particular peaces of info a lot of people want, then new functions can be written to retreave this info.

Although the code is here, it is not enabled in NVDA by default yet as server dll will need to be included in misc deps.