Installer changes to solve NVDA freezing when the installer launches the uninstaller. Specifically: *Rather than a message box alerting the user that a previous install must be uninstalled, instead show an actual installer page which reports this information, and also provides a check box which allows the user to decide whether the previous install should be uninstalled or not. It is checked by default. This also gives the user a chance to cancel the install if they are not happy uninstalling the previous copy and also no longer wishes to go further with this install. *No longer launch the uninstaller from the installer with the silent switch, but instead provide it with a /nonInteractive switch. The /nonInteractive switch is not standard to nsis so older uninstallers will still ask the user to confirm the uninstall, and also may show a reboot question at the end. *Support the /nonInteractive switch in the uninstaller by skipping the confirm and finnish pages, so just only showing the instfiles page. This means that for all new uninstallers, run from the installer, all the user will see is a dialog showing the uninstaller's progress, and the user will not have to answer any questions. *Removed some no longer used installer language strings.
The end result of all these changes is that when the user runs the installer, if a previous install is detected, the user will have the choice of uninstalling the old copy (which is strongly recommended). If an older copy (older than this revision) is being uninstalled, the user will have to go through the uninstaller in full, answering the confirm and finnish pages (though the user can safely decide not to reboot at this point). However if a copy of NVDA of this revision or newer is being uninstalled, then all the user will see is an uninstall progress dialog, and just has to wait until its complete. Previously Either NVDA would freeze when it started the uninstaller, or focus would seem to drop back to what ever was behind the installer.
Translators: please note that two language strings have been added to the installer.