Initial implementation of dynamic NVDAObject classes.
When an NVDAObject is instanciated, it always creates an object with a type that is a subclass of the class used for instanciation, however it may also contain other suitable base classes depending on the arguments given to the object.
For example: When instanciating a Window NVDAObject, if its window class is 'Edit' then you will actually be creating an object who's type is based on Edit Window. 
Or, if for instance you instanciate an IAccessible NVDAObject, which happens to have a window class of EXCEL7, then you will really create an object who's type is based on IAccessible and ExcelGrid window.

The base NVDAObject's __new__ method takes care of this by calling the class's findBestClass class method, which returns the most suitable subclass for that class, based on the arguments given. Implementations of findBestClass may call super, meaning that lower base classes also get a chance to select the most suitable subclass.

So far a findBestClass method has been provided for NVDAObject, NVDAObjects.window.Window and NVDAObjects.IAccessible.IAccessible.

NVDAObjects fully compatible with Window have been moved in to the NVDAObjects.window package.

findBestClass should only select from classes in its class's own package. classes in the same package should all be compatible (i.e. they take the same arguments for instanciation).

A lot of work still needs to be done to make NVDA provide the same user experience as it did before this new implementation. many IAccessible NVDAObject subclasses are not yet mapped in IAccessible NVDAObject's findBestClass method.

However, things like edit, richEdit, cintilla, winConsole and winword are mapped in Window's findBestClass method, so these should work ok.