Better implementation for self-voicing applications. Specifically: *appModules can now set a class variable called 'selfVoicing' to True if the application is completely self-voicing. *eventHandler.executeEvent_appModuleLevel and scriptHandler.findScript_appModuleLevel now drop any requests if the appModule they fetch has selfVoicing set to True. *mouseHandler now uses NVDA's event mechanism to execute event_mouseMoved. *Added an event_typedCharacter to the base NVDAObject, and NVDAHelper queues a 'typedCharacter' event, rather than queuing speakTypedCharacters directly. *eventHandler.queueEvent now can take optional keyword arguments -- executeEvent has been able to for a while.
This implementation is much better than the old way of simply toggling speech to off. A part from it being possible for an appModule developer to forget to toggle speech back on, this new implementation also ignores *all* events and scripts, so performance won't be needlessly slowed down as much.

Note that it is still possible to write the old type of appModule, perhaps if Braille was still needed. Though for action games etc, where not even Braille would be useful, this is the much cleaner solution.