nvdaHelperRemote changes: *Rather than using VBufContainer, VBufRemote now loads a backend dll and uses the dll's createBackend function, and then uses the created backend as the buffer for all of the rpc functions.  So it can eventually free the backend dll, it keeps a map of active backends to dll handles. *rpcSrv: change the order of some code so it can compile again. *Initialize and terminate rpcSrv in inProcess_initialize and inProcess_terminate. *Compile in rpcSrv and VBufRemote when building nvdaHelperRemote.dll. *Fix a typo in rpcSrv.h. *Export the hook registration functions from nvdaHelperRemote.dll as the VBufBackend dlls may want to make use of them.
These changes finally mean that nvdaHelperRemote now contains an rpc server, plus an implementation of the VBuf interface.

Note that None of this code has been tested, though it does seem to compile ok. VBufLib and the backends still need to be worked on further, and properly integrated in to the nvdaHelper build system.