*Convert MVDAObjects.IAccessible.MSHTML.MSHTML in to a real APIClass by implementing kwargsFromSuper. *NVDAObjects.IAccessible.IAccessible: implement a getPossibleAPIClasses method, which yields MSHTML NVDAObject if IAccessibleChildID is 0 or not in kwargs at all. *Remove selection of MSHTML out of NVDAObject.IAccessible.IAccessible's findOverlayClasses as its now selected as a possible APIClass.
The advantage of this change is that  now MSHTML  is used  only if it really is supported by the IAccessible object (i.e. we can queryService to it), rather than it being used simply if the window class was 'Internet Explorer_Server'.
This does mean that  we always try queryService a lot more than we did, but it doesn't seem too slow.