UIAHandler: 	* Merge the UIAEventListener class into the UIAHandler class. This stops us from having to keep and dereferencing a weakref to the UIAHandler instance from the event listener. One class is also cleaner. 	* No longer dynamically generate the UIAutomation typelib interfaces on import, instead just rely on NVDA already having them in comInterfaces. This certainly is now always true, though may not have been during Windows 7 betas when UIAutomation was still changing. 	* As we can now assume we always have the UIAutomation COM interfaces, no longer condissionaly define all our UIA-to-NVDA mappings and classes. Instead just define them no matter what. Takes out one level of indenting. 	* Create all needed global UIAutomation COM object instances (IUIAutomation, base cache requests, treeWarkers etc) and register UIAutomation events, in an MTA thread. 	* Finally allow NVDA to unregister UIAutomation events on exit, inside the MTA thread. This used to cause a freeze if done in our main thread so we never bothered.
having events now in an MTA may speed up NVDA a touch as the events will no longer be interupting our main thread. As we now cexplicitly unregister all UIAutomation events on exit, this may stop Windows explorer becoming very sluggish after hours of use and NVDA restarts (though not sure). And finally as we now register for UIAutomation events in an MTA, this will stop the freeze at NVDA start up that occures on Windows 7 if another app has crashed and its crash dialog is still up. (UIAutomation tried calling NtUserGetComboBoxInfo on the crashed window and it does not timeout).