Fix for #574 (Freeze when opening Windows Live Mail Move to and Copy to dialogs) plus any other IAccessible implementation that may use ROLE_SYSTEM_WINDOW on a client or other non-windowRoot object: 	* Move the special code in IAccessible NVDAObject's next, previous and parent properties that did something specific for the windowRoot (i.e. used Window NVDAObject's navigation properties instead) to newly implemented navigation properties on the WindowRoot IAccessible NVDAObject class. 	* IAccessible NVDAObject's children property: when checking if a child is a WindowRoot IAccessible object, instead of just using accRole, instead use IAccIdentity to check for an objectID of 0 and childID of 0. 	* IAccessible NVDAObject's findOverlayClasses:  only choose to use WindowRoot if the objectID is 0 (previously it checked for either 0 or None).
These changes make sure that WindowRoot code is only used for real windowRoot IAccessibles provided by oleacc, as opposed to also any IAccessible that happened to have a role of window.
In the Windows Live Mail dialog case: there were one or more client IAccessibles with an accRole of ROLE_SYSTEM_WINDOW and therefore object navigation was getting confused and looping back to a parents next/previous, and getDialogText was going in to endless recursion as children kept returning children of that window's client.