When updating a previous installation, old eSpeak voices are now removed.
Before copying in the new files, everything in the espeak-data\voices directory is removed except for variants.
Otherwise, multiple copies of voices and other files can be left around when they've been moved in eSpeak. In particular, the recent reorganisation of espeak-data\voices means that there are a lot of duplicates.
Fixes #3181.