scons: Building eSpeak and liblouis is now manditory; the build will fail if they don't exist.
Previously, this was optional to enable a smooth transition from when we distributed pre-built eSpeak and liblouis dependencies. However, we haven't done this for a long time and won't be doing it again.
This prevents broken (but apparently successful) builds if eSpeak or liblouis aren't in the correct place on the build system.
Re #3195.