NVDA no longer fails to start if the system time is earlier than the last check for an update.
When this happens (e.g. to a CMOS battery change), the time since the last check could be a very large negative number, which in turn becomes a large positive number when calculating the time to the next check. If it's large enough, this will cause an overflow.
Therefore, restrict the time since the last check so it can never be earlier than now (must be >= 0).
Fixes #3260.