Adding and deleting config profiles take effect immediately, so changes to triggers must do likewise.
Previously, triggers were stored in the main configuration, which might not be saved on exit or might be reverted.
Triggers are now saved in their own file so they can be saved immediately without forcing the rest of the config to be saved.
Unfortunately, this will cause any existing triggers to be lost.