When adding a gesture with multiple possible identifiers, honour the user's selection instead of always using the last identifier for any selection.
Oops. Clearly didn't test that too well. :(
This occurred because a lambda is used for the EVT_MENU events, but the lambda will always use the value of any variables at the time it is called. Classic Python newbie error. :( Instead, pass the value we actually want as a default keyword argument so it is determined when the lambda is *created*.