PowerPoint appModule: Again try to support PowerPoint 2010 protected mode, but keeping support for PowerPoint 2013 slideShows. Also stops PowerPoint complaining in 2013 protected mode, but therefore no support for 2013 protected mode.
Specifically: Go back to a structure that uses a fetchPpObjectModel method in chooseNVDAObjectOverlayClasses rather than using a ppApplication property. However, fetchPpObjectModel (and its helper methods) are now smarter in that they try various things such as fetching the object model from the window (if a window is provided and its safe to do so), or getActiveObject via rpc etc. As there is now no ppApplication property, move ppVersionMajor property off the appModule and onto the PaneclassDC NVDAObject.
So, the code should work in the following ways:
 * PowerPoint 2003 to 2010 normal and protected modes (including normal  and slideShow views): the object model wil be fetched from the window with AccessibleObjectFromWindow.
  * PowerPoint 2013 protected mode (normal view): not supported at all. wm_getObject causes PowerPoint to complain about stability. GetActiveObject via rpc or otherwize gives us an application with no active presentation.
   * PowerPoint 2013 normal mode normal view, and slideShow view (normal and protected mode): getActiveObject via rpc is used.