Fix some cases where a profile trigger isn't exited when an application closes.
appModuleHandler.update is called from IAccessibleHandler for all events, but it currently cleans up app modules for dead processes. Sometimes, an event is fired before we fetch the old app modules in api.setFocusObject, so by that point, the app module is gone and we can't exit the trigger.
To fix this, split the cleanup code into a separate function and call it only after the old app modules have been fetched in setFocusObject.
Re #3732.