Fix wx locale initialisation to prevent "locale 'xxx' cannot be set" error.
We were passing the string locale identifier as the first argument, but we're actually supposed to pass a numeric wx language identifier. wxPython 2.8 didn't seem to complain about this, but wxPython 3 presents an error dialog on startup.
Instead, use wx.Locale.FindLanguageInfo to get the wx language info from the locale identifier and pass the wx language identifier from that.