watchdog: Only wake up if action might actually be required rather than checking using a continual timer.
A waitable timer is used to achieve this. When the core reports as alive, the timer is set so that the watcher will wake if the timer isn't reset or cancelled.
Also, the core pump reports that it is going to sleep at the end of the pump, which cancels the timer. This is necessary because the core doesn't spin any more, so if it isn't doing anything, it won't report as alive.