Fix an issue where when opening a message in Outlook Express, the message sometimes wouldn't get focused in NVDA until after a key was pressed, causing apparent double speaking in some cases.
Wx timers aren't re-entrant, so if a modal loop runs, the timer just gets ignored. Unfortunately, sending a key press causes a modal loop, so the pump request for the focus event was sometimes being swallowed.
Now, before it exits, the pump checks whether there is a pending request but a pump isn't scheduled. If this occurs, it schedules another one.
Re #3803.