Fix an issue where two copies of NVDA sometimes started after logging on in Windows 8 or later.
This occurred because EoA seems to start NVDA on the Default desktop before Windows switches to it, but NVDA uses the name of the current input desktop for its startup mutex. Worse, NVDA didn't have privileges to open the input desktop (probably still Winlogon), so the name was empty.
Instead, NVDA now uses the name of the desktop for its thread, which is definitely more correct.