Fix an issue where focus would be moved out of the Start Screen if NVDA started after logon on Windows 8 and later.
This occurred because NVDA shows and hides its window briefly so that Windows will allow it to come into the foreground on demand later. Unfortunately, this can also mess with the focus.
When we have uiAccess, this isn't necessary, as uiAccess applications can always come to the foreground, so don't do this if we have uiAccess.