cancellableSendMessageTimeout: Don't ever return early due to cancellation while the background SendMessage is still running, as this can cause crashes in the remote process.
Sometimes, the caller allocates memory in the remote process, sends a message which depends on it and frees it as soon as the SendMessage* call returns. If the background SendMessageTimeout is still running, the message might reach the remote process and it will then use the freed memory, thus causing crashes or heap corruption.
To fix this, the background thread attempts to send the message with 400ms timeouts, checking for cancellation after each attempt, similar to what the main thread did previously. The main thread just waits for the background thread to finish handling the message, regardless of whether it was cancelled. For the most part, this takes us back to what we had before the background thread was implemented, except that we don't get blocked in SendMessageTimeout for longer than we should.
This will hopefully fix crashes in Outlook Express and other applications.
Re #3859.