cancellableSendMessageTimeout: When sending a message from a thread other than the main one, even though we don't use a background thread, still send in short attempts and check for cancellation.
UI Automation threads send messages to potentially unresponsive windows. This allows us to check for cancellation, though because cancellation is based on the main thread being alive, messages might not be cancelled if the main thread recovers before the thread in question. However, more importantly, sending in short attempts allows us to get around the issue where SMTO_ABORTIFHUNG doesn't abort if the window hangs while the operation is in progress instead of before.