Call chooseNVDAObjectOverlayClasses on an AppModule even when it is implemented in a superclass and not overridden.
Previously, it was only called if it was directly implemented by the class. The only case it isn't called now is if the base implementation hasn't been overridden, as it does nothing, so this call is a complete waste.
This allows an ugly hack to be removed from the new version of the totalcmd64 app module.