Fix UIA implementations which use IAccessibleEx such as the text pattern in Internet Explorer.
The UIA client requests a private interface on the factory entry object. If it doesn't get this, IAccessibleEx implementations don't work. Get around this by forwarding QueryInterface on our overridden factory entry to the original factory entry.
Unfortunately, when we do this, our proxy doesn't seem to be used except in very rare cases, most likely only when the real proxy can't go in-process. This makes this pretty useless.