Initial support for the new version of MathPlayer which is a library independent of IE.
Users can interact with math under the review cursor by pressing NVDA+alt+m.
To support this, an NVDAObject or TreeInterceptor must implement a getMathMlForEquation method which takes a TextInfo and returns the MathML for the math therein. It can raise ValueError if this isn't math.
This isn't implemented anywhere yet. Implementations to follow.