Revert t3800 (removal of most UI Automation client-side proxies).
This was causing freezes in Chicken Nugget and possibly other issues. This was probably due to the MSAA proxy being used for controls for which it isn't normally used.
This idea doesn't really solve what it was intended to solve anyway, as we can't stop the MSAA proxy from being used when we don't use UIA without breaking cases where it is needed.
Reverts e1abd8aa5 388d0b3f 59f3211f.
Fixes #4058. Re #3800.