Fix incorrect naming of the Microsoft Word document control as Clipboard when Word starts and probably other similar issues.
This was triggered by an incorrect focus event on an SDM control. When we calculate the SDM child, we use the focus window, but in the case of these incorrect events, the focus window is now something entirely unrelated.
To fix this, when calculating the SDM child, ensure the focus window is a descendant of the SDM window.