Fix continual announcement of Mozilla (and perhaps other) editable areas when typing in Korean and other east Asian languages.
Caret events are fired whenever a character is typed. However, when input composition is in progress, we fake the focus. The caret event was causing the focus to be bounced out of the input composition object back to the editable area every time a character was typed.
Fix this by ignoring these caret events when the focus is not an IAccessible.
Re #4244.