Fix incorrect breaking of words at marks such as vowel signs and virama in Indic languages.
speech.speakTypedCharacters and textInfos.offsets.find{Start,End}OfWord were using unicode.isalnum to check for characters that are part of a word, but this only covers alphanumeric characters. The marks in question aren't alphanumeric, but should still be considered part of a word.
Therefore, use the Unicode category of the character and include letters, marks and numbers.
Re #4254.