Provide framework to support loading specific app modules for executables which host many different apps.
For example, this will allow specific app modules to be loaded for apps hosted by javaw.exe, rather than always loading just the javaw app module.
This is done by implementing the getAppNameFromHost function in the module for the hosting executable (not the AppModule class). See the code documentation for the AppModule class for details.
Note that this commit doesn't include any modules which actually implement this.