Back out t4294.
This is at least a partial cause of #4381. I think this is due to the removal of wx.Yield in KeyboardInputGesture.send. If there is pending stuff in the queue, this would make the first wx.Yield indirectly called by EditableText._hasCaretMoved take longer, so waiting for caret movement would time out sooner than it did before. Technically, this isn't a bug in #4294.
However, we have no current need for #4294 any more and we've had reports that suggest it might be causing modifiers to stick. Therefore, until we have time to debug this or until there is a specific need for it, it's simpler to just back it out.
This reverts commits 6fb9c104..45bb5962.
Re #4294, #4381.