Ensure that the NVDA Notification Area icon gets cleaned up properly.
We now seem to need another iteration of the main loop in order for the TaskBarIcon to be destroyed properly. The documentation does seem to suggest that Destroy runs in a subsequent loop iteration, but we didn't need this before wxPython 3 for some reason.
Unfortunately, the loop has already exited by the time we destroy things. ProcessPendingEvents doesn't seem to work; Pending() even reports False after the destroy. Thankfully, calling MainLoop again does work and it returns quickly because the top window has been destroyed.
Re #4460.