NVDA now automatically reports new text in mintty.
mintty is a fork of PuTTY. For the most part, it behaves the same, even as far as WM_CHAR not working. In order to benefit from any future enhancements made to the PuTTY app module, the mintty app module inherits from it. However, the window class is different, so the PuTTY app module now allows this to be overridden.