Split table cells into a new TextInfo presentation category (PRESCAT_CELL), as they are quite different to other PRESCAT_MARKERs.
For the other markers, the control field text is generally more interesting than the content itself (which is usually just a space). This is not the case for table cells, which are the same as containers except we don't want to report when they end.
Also, math is now PRESCAT_MARKER, not PRESCAT_SINGLELINE.