In Skype for Desktop 7.2 and later, typing notifications are now reported and problems immediately after moving focus out of a conversation have been fixed.
Skype 7.2 changed the typing notification control. Aside from the obvious, this also caused an exception whenever a conversation lost focus.
Also, since Skype seem intent on changing things every couple of versions, if the output list and/or typing indicator can't be found, log a warning and try to handle this more gracefully.