When using text expansion functionality in software such as AutoHotkey, expanded text is no longer garbled when NVDA is configured to handle keys from other applications.
This occurred because we ignore all injected keys while we're injecting, but that might also ignore injected keys from other apps that happen to arrive at the same time. Instead, to ignore just our own injected keys, we now set the dwExtraInfo parameter to our process id when injecting and check for that in the hook.