In Windows 7 and later, text is no longer garbled when typing in certain applications such as Wordpad with a braille display and NVDA's Announce formatting changes after the cursor setting enabled.
This is related to Rich Edit controls. It occurred because in order ot get formatting information in these controls, we have to move the selection (and thus the cursor), which obviously causes problems if typed characters are arriving during this process.
To fix this, we can use ITextDocument. We previously disabled this due to poor performance, but it seems this is fixed in Windows 7 and later. Therefore, use ITextDocument in Windows 7 and later.