When an application is responding slowly and you switch away from that application, NVDA is now much more responsive in other applications in most cases.
This is done by filtering out events from background windows in most cases to avoid freezes when a background app is unresponsive. One hard-coded exception is when background progress bar reporting is enabled, in which case background valueChange events are allowed.
To facilitate this, eventHandler.shouldAcceptEvent was introduced.
eventHandler.requestEvents was also added to request particular events that are blocked by default; e.g. show events from a specific control or certain events even when in the background.
Fixes #3831.