Really fix garbled text in Wordpad, etc. with a braille display.
Our ITextDocument support was still moving the selection to determine whether a chunk of text was a link, so it actually had the same problem as before with reporting of links enabled.
We now duplicate the range and use the tomLink unit to determine whether a chunk is a link.