Windows 10 will no longer repeatedly warn that NVDA is incompatible. NVDA now functions on the sign-in screen as expected in Windows 10 build 10525.
The Windows compatibility block should be removed in a Windows Update any day now, but it looks like this may not be available in time for NVDA 2015.3. Aside from the annoying notifications, Windows build 10525 seems to stop apps with compatibility blocks from running on the sign-in screen.
To work around this, the publisher has been changed for official builds so that the compatibility block doesn't match.
The real change is actually on the build server, not in NVDA itself. This is just an empty commit to denote the change.
Fixes #5253, #5303.