Filter out events from UIA elements which aren't in the foreground even where the element isn't a window element.
Previously, eventHandler.shouldAcceptEvent was only called if the element's NativeWindowHandle was valid. However, this is only ever valid for window elements. We now use the nearest ancestor window handle in this case.
To do this, querying of the element's CachedNativeWindowHandle is now handled by getNearestWindowHandle. This also simplifies the code somewhat.
Note that we were always fetching the nearest window handle for isNativeUIAElement anyway (which gets called before filtering), so this shouldn't result in extra queries.