App module handler: Executables with dots as file name are now loaded by replacing dots with underscores. re #5323.
Because of Python's way of importing modules, an executable name such as foo.bar would be interpreted as bar module in foo package. To handle this case, underscores will be used. This may also mean how app module is reported to users (Control+NVDA+F1) may have to be modified somehow. For now, each dots will be replaced bd a single underscore.