Remove incorrect vmIDsToWindowHandles cache
The vmIDsToWindowHandles cache is base on an incorrect assumption
that there is a mapping between a jvmid to a _single_ Window Handle.
This is incorrect as a Java application can have many top level
windows (each top level frame corresponds to a unique Window Handle).

The cache was introduced in 2007:

https://github.com/nvaccess/nvda/commit/d4a74c7cf155cedcd1cf48b47604a9
c5d667e83b

This code was incorrect at the time, but was good enough to make things
work with Java applications that have only one top level frame.

A commit in 2010 implemented the correct way of retrieving the window
handle from a given AccessibleContext:

https://github.com/nvaccess/nvda/commit/eddfb4bd3e193ae4145dd6a697babc
4bc12582e5

This effectively removed the need of using cache, but the cache was
not removed for some reason.

The symptom of the issue is described in bug #5732.