Add-ons subsystem: Base routines to disable add-ons upon request. re #3090.
In order to allow users to disable add-ons from add-ons manager or from command line, it is necessary to tweak the add-ons subsystem a bit.
Specifically:
* Core.py: If all add-ons are disabled, this fact is recorded in the log, useful for troubleshooting purposes.
* addonHandler.py: Added a new pending disable set so add-ons can be disabled upon startup. The user interface to manipulate this set will be committed later. Also added a function (disableAddonsIfAny) that'll save current disable state to a private set (can be extended to supporting adding add-ons from command line).
* addonHandler.Addon: Added isDisabled property to report if the currently selected add-on is disabled. Also modified isRunning attribute to check if the add-on is disabled. These properties will come in handy when presenting state information in add-ons manager.
So far, tests show it is working. Next on the list is adding few more properties such as pending enable/disable to addonHandler.Addon, then start implementing the user interface.