Add-ons subsystem, UI and user guide: enable/disable status are now shown, documented add-on enable/disable feature. re #3090.
Although it was suggested to use a dedicated set for pending enables, it is better to use a module-level set, as pendingDisableSet is really intended to tell disabled add-ons set what to disable. Also, because isDisabled flag takes precedence, enable/disalbe status is fetched early.
Also documented add-on enable/disable feature in the user guide. As of this commit, all the pieces for enabling/disabling add-ons and its user interface are there (next up is command-line version, a bit tricky but doable).